Skeleton pygame window
Created main.py

handle_event_quit, captures the quit
event when the user hits the x box
on the window and closes. This function
is used as the condition in the main
while loop of the program.

draw window, paints window(WIN)
with WHITE color.

Created package gol_modules and within
gol_params.py. This will have all constants to be used in the game.